Convert storm to wheels
