+++ .travis.yml
@@ -74,6 +74,8 @@ jobs:
     - stage: test
       env: CHECK=riak_repl PYTHON3=true
     - stage: test
+      env: CHECK=storm PYTHON3=true
+    - stage: test
       env: CHECK=traefik PYTHON3=true
     - stage: test
       env: CHECK=upsc PYTHON3=true
