--- storm/conf.yaml.example
@@ -1,25 +1,0 @@
-init_config:
-  # Default environment to use for all instances
-  # environment: production
-  #
-  # Intervals (in seconds): Default is 1 min
-  # intervals:
-  #   - 60
-
-instances:
-  # - server: http://localhost:8080
-  #   # Always specify the server with the default storm REST api port.
-  #   # Default is 8080
-  #
-  #   # Specify the environment
-  #   environment: preproduction
-  #
-  #   #Specify additional tags
-  #   tags:
-  #     - mytag:myvalue
-  #     - mytag2:myvalue2
-  #
-  #   # specify metric intervals in seconds (note this overrides the defaults).
-  #   intervals:
-  #     - 60
-  #
