+++ storm/datadog_checks/storm/__init__.py
@@ -1,0 +1,7 @@
+from .__about__ import __version__
+from .storm import StormCheck
+
+__all__ = [
+    '__version__',
+    'StormCheck'
+]
