+++ storm/datadog_checks/storm/data/conf.yaml.example
@@ -1,0 +1,25 @@
+init_config:
+  # Default environment to use for all instances
+  # environment: production
+  #
+  # Intervals (in seconds): Default is 1 min
+  # intervals:
+  #   - 60
+
+instances:
+  - server: http://localhost:8080
+  #  Always specify the server with the default storm REST api port.
+  #  Default is 8080
+  #
+  #  Specify the environment
+  #  environment: preproduction
+  #
+  #  Specify additional tags
+  #  tags:
+  #    - mytag:myvalue
+  #    - mytag2:myvalue2
+  #
+  #   Specify metric intervals in seconds (note this overrides the defaults).
+  #   intervals:
+  #     - 60
+  #
