+++ storm/logos/README.md
@@ -1,0 +1,19 @@
+# Logos
+
+The logos folder should contain three images with filenames and sizes that exactly match the following specifications. Underneath each specification is a list of places where the images may appear in the web app.
+
+#### `saas_logos-bot.png` (200 × 128)
+
+- Integration tile images at `/account/settings`
+- Description heading at `/account/settings#integrations/storm`
+- Integration monitor tiles and search bar results images at `/monitors#create/integration`
+
+#### `saas_logos-small.png` (120 × 60)
+
+- Integration dashboards list images at `/dash/list`
+- Some integration dashboards/screenboards at `/dash/integration/storm`
+
+#### `avatars-bot.png` (128 × 128)
+
+- Event stream at `/event/stream`
+- Notification icons at `/report/monitor`
