+++ storm/manifest.json
@@ -13,5 +13,6 @@
   "creates_events": false,
   "type": "check",
   "is_public": true,
-  "metric_prefix": "storm."
+  "metric_prefix": "storm.",
+  "integration_id": "storm"
 }
