+++ storm/requirements-dev.txt
@@ -1,0 +1,2 @@
+datadog-checks-dev
+responses
