--- storm/requirements.txt
@@ -1,1 +1,0 @@
-# integration pip requirements
