+++ storm/tests/common.py
@@ -1,0 +1,597 @@
+# (C) Datadog, Inc. 2010-2016
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+# flake8: noqa E501
+
+INSTANCE = {
+    'server': 'http://localhost:8080',
+    'environment': 'test'
+}
+
+TEST_STORM_CLUSTER_SUMMARY = {
+    "executorsTotal": 33,
+    "stormVersion": "1.0.3",
+    "slotsTotal": 10,
+    "slotsFree": 4,
+    "user": None,
+    "topologies": 1,
+    "supervisors": 1,
+    "central-log-url": None,
+    "bugtracker-url": None,
+    "tasksTotal": 33,
+    "slotsUsed": 6
+}
+
+TEST_STORM_NIMBUSES_SUMMARY = {
+    "nimbuses": [
+        {
+            "nimbusUptimeSeconds": "Not applicable",
+            "nimbusUpTime": "Not applicable",
+            "version": "Not applicable",
+            "status": "Offline",
+            "nimbusLogLink": "http://nimbus01.example.com:9006/daemonlog?file=nimbus.log",
+            "port": "6627",
+            "host": "nimbus01.example.com"
+        },
+        {
+            "nimbusUpTimeSeconds": 25842,
+            "nimbusUpTime": "7h 10m 42s",
+            "version": "1.0.3",
+            "status": "Leader",
+            "nimbusLogLink": "http://1.2.3.4:9006/daemonlog?file=nimbus.log",
+            "port": 6627,
+            "host": "1.2.3.4"
+        }
+    ]
+}
+
+TEST_STORM_SUPERVISOR_SUMMARY = {
+    "logviewerPort": 9006,
+    "schedulerDisplayResource": False,
+    "supervisors": [
+        {
+            "uptimeSeconds": 31559,
+            "slotsTotal": 10,
+            "version": "1.0.3",
+            "slotsUsed": 6,
+            "totalMem": 3072,
+            "host": "1.2.3.4",
+            "id": "11111111-1111-1111-1111-111111111111",
+            "uptime": "8h 45m 59s",
+            "totalCpu": 900,
+            "usedCpu": 0,
+            "logLink": "http://1.2.3.4:9006/daemonlog?file=supervisor.log",
+            "usedMem": 4992
+        }
+    ]
+}
+
+TEST_STORM_TOPOLOGY_SUMMARY = {
+    "schedulerDisplayResource": False,
+    "topologies": [
+        {
+            "requestedTotalMem": 0,
+            "assignedMemOffHeap": 0,
+            "assignedCpu": 0,
+            "uptimeSeconds": 1525505,
+            "schedulerInfo": None,
+            "uptime": "17d 15h 45m 5s",
+            "id": "my_topology-1-1489183263",
+            "assignedMemOnHeap": 4992,
+            "encodedId": "my_topology-1-1489183263",
+            "requestedMemOnHeap": 0,
+            "owner": "storm",
+            "assignedTotalMem": 4992,
+            "name": "my_topology",
+            "workersTotal": 6,
+            "status": "ACTIVE",
+            "requestedMemOffHeap": 0,
+            "tasksTotal": 33,
+            "requestedCpu": 0,
+            "replicationCount": 1,
+            "executorsTotal": 33
+        }
+    ]
+}
+
+TEST_STORM_TOPOLOGY_RESP = {
+    "assignedMemOffHeap": 0,
+    "topologyStats": [
+        {
+            "failed": None,
+            "acked": 104673,
+            "completeLatency": "285.950",
+            "transferred": 307606,
+            "emitted": 307606,
+            "window": ":all-time",
+            "windowPretty": "All time"
+        }
+    ],
+    "assignedCpu": 0,
+    "uptimeSeconds": 1525788,
+    "executorsTotal": 33,
+    "bolts": [
+        {
+            "errorWorkerLogLink": "",
+            "capacity": "0.000",
+            "processLatency": "201.474",
+            "executors": 3,
+            "boltId": "Bolt1",
+            "failed": 0,
+            "errorHost": "",
+            "tasks": 3,
+            "errorTime": None,
+            "emitted": 101309,
+            "executeLatency": "0.001",
+            "transferred": 101309,
+            "errorPort": "",
+            "errorLapsedSecs": None,
+            "acked": 212282,
+            "encodedBoltId": "Bolt1",
+            "lastError": "",
+            "executed": 106311
+        },
+        {
+            "errorWorkerLogLink": "",
+            "capacity": "0.000",
+            "processLatency": "0.010",
+            "executors": 2,
+            "boltId": "Bolt2",
+            "failed": 0,
+            "errorHost": "",
+            "tasks": 2,
+            "errorTime": None,
+            "emitted": 0,
+            "executeLatency": "0.015",
+            "transferred": 0,
+            "errorPort": "",
+            "errorLapsedSecs": None,
+            "acked": 3153,
+            "encodedBoltId": "Bolt2",
+            "lastError": "",
+            "executed": 3153
+        },
+        {
+            "errorWorkerLogLink": "",
+            "capacity": "0.000",
+            "processLatency": "0.003",
+            "executors": 3,
+            "boltId": "Bolt3",
+            "failed": 0,
+            "errorHost": "",
+            "tasks": 3,
+            "errorTime": None,
+            "emitted": 0,
+            "executeLatency": "0.009",
+            "transferred": 0,
+            "errorPort": "",
+            "errorLapsedSecs": None,
+            "acked": 4704,
+            "encodedBoltId": "Bolt3",
+            "lastError": "",
+            "executed": 4704
+        },
+        {
+            "errorWorkerLogLink": "",
+            "capacity": "0.000",
+            "processLatency": "291.756",
+            "executors": 4,
+            "boltId": "Bolt4",
+            "failed": 0,
+            "errorHost": "",
+            "tasks": 4,
+            "errorTime": None,
+            "emitted": 101607,
+            "executeLatency": "0.001",
+            "transferred": 101607,
+            "errorPort": "",
+            "errorLapsedSecs": None,
+            "acked": 218808,
+            "encodedBoltId": "Bolt4",
+            "lastError": "",
+            "executed": 110946
+        },
+        {
+            "errorWorkerLogLink": "",
+            "capacity": "0.000",
+            "processLatency": "1014.634",
+            "executors": 2,
+            "boltId": "Bolt5",
+            "failed": 0,
+            "errorHost": "",
+            "tasks": 2,
+            "errorTime": None,
+            "emitted": 17,
+            "executeLatency": "0.001",
+            "transferred": 17,
+            "errorPort": "",
+            "errorLapsedSecs": None,
+            "acked": 208890,
+            "encodedBoltId": "Bolt5",
+            "lastError": "",
+            "executed": 104445
+        },
+        {
+            "errorWorkerLogLink": "",
+            "capacity": "0.000",
+            "processLatency": "0.005",
+            "executors": 3,
+            "boltId": "Bolt6",
+            "failed": 0,
+            "errorHost": "",
+            "tasks": 3,
+            "errorTime": None,
+            "emitted": 0,
+            "executeLatency": "0.010",
+            "transferred": 0,
+            "errorPort": "",
+            "errorLapsedSecs": None,
+            "acked": 4705,
+            "encodedBoltId": "Bolt6",
+            "lastError": "",
+            "executed": 4705
+        }
+    ],
+    "schedulerDisplayResource": False,
+    "replicationCount": 1,
+    "requestedCpu": 0,
+    "tasksTotal": 33,
+    "visualizationTable": [],
+    "debug": False,
+    "requestedMemOffHeap": 0,
+    "spouts": [
+        {
+            "errorWorkerLogLink": "http://1.2.3.4:9006/log?file=my_topology-1-1489183263%2F6707%2Fworker.log",
+            "lastError": "com.rabbitmq.client.ShutdownSignalException: clean connection shutdown; protocol method: #method<connection.close>(reply-code=200, reply-text=OK, class-id=0, method-id=0)\n\tat com.rabbitmq.client.impl.",
+            "acked": 104673,
+            "errorLapsedSecs": 38737,
+            "errorPort": 6707,
+            "transferred": 104673,
+            "encodedSpoutId": "source",
+            "emitted": 104673,
+            "spoutId": "source",
+            "errorTime": 1490670314,
+            "tasks": 8,
+            "errorHost": "1.2.3.4",
+            "failed": 0,
+            "completeLatency": "285.950",
+            "executors": 8
+        }
+    ],
+    "status": "ACTIVE",
+    "user": None,
+    "msgTimeout": 300,
+    "windowHint": "All time",
+    "encodedId": "my_topology-1-1489183263",
+    "requestedMemOnHeap": 0,
+    "owner": "storm",
+    "window": ":all-time",
+    "assignedTotalMem": 4992,
+    "samplingPct": 10,
+    "assignedMemOnHeap": 4992,
+    "id": "my_topology-1-1489183263",
+    "configuration": {
+        "drpc.request.timeout.secs": 600,
+        "storm.auth.simple-acl.users.commands": [],
+        "nimbus.thrift.max_buffer_size": 1048576,
+        "logviewer.appender.name": "A1",
+        "storm.messaging.netty.transfer.batch.size": 262144,
+        "storm.exhibitor.poll.uripath": "/exhibitor/v1/cluster/list",
+        "topology.name": "my_topology",
+        "storm.id": "my_topology-1-1489183263",
+        "topology.kryo.decorators": [],
+        "ui.port": 9005,
+        "java.library.path": "/usr/local/lib:/opt/local/lib:/usr/lib",
+        "drpc.invocations.threads": 64,
+        "storm.auth.simple-acl.users": [],
+        "topology.trident.batch.emit.interval.millis": 500,
+        "storm.nimbus.retry.intervalceiling.millis": 60000,
+        "topology.disruptor.wait.timeout.millis": 1000,
+        "topology.min.replication.count": 1,
+        "ui.header.buffer.bytes": 4096,
+        "ui.filter": None,
+        "backpressure.disruptor.high.watermark": 0.9,
+        "ui.http.x-frame-options": "DENY",
+        "topology.worker.max.heap.size.mb": 1024,
+        "supervisor.childopts": "-Xmx256m",
+        "client.blobstore.class": "org.apache.storm.blobstore.NimbusBlobStore",
+        "storm.blobstore.acl.validation.enabled": False,
+        "storm.zookeeper.auth.password": None,
+        "supervisor.worker.timeout.secs": 30,
+        "transactional.zookeeper.servers": None,
+        "ui.users": None,
+        "pacemaker.childopts": "-Xmx1024m",
+        "logviewer.max.sum.worker.logs.size.mb": 4096,
+        "worker.heap.memory.mb": 768,
+        "storm.blobstore.replication.factor": 3,
+        "nimbus.cleanup.inbox.freq.secs": 600,
+        "nimbus.queue.size": 100000,
+        "nimbus.seeds": [
+            "nimbus01.example.com"
+        ],
+        "nimbus.topology.validator": "org.apache.storm.nimbus.DefaultTopologyValidator",
+        "worker.gc.childopts": "",
+        "topology.kryo.register": None,
+        "topology.kryo.factory": "org.apache.storm.serialization.DefaultKryoFactory",
+        "topology.component.resources.onheap.memory.mb": 256,
+        "storm.messaging.netty.authentication": False,
+        "topology.disable.loadaware.messaging": False,
+        "storm.messaging.transport": "org.apache.storm.messaging.netty.Context",
+        "topology.error.throttle.interval.secs": 10,
+        "drpc.http.port": 3774,
+        "topology.component.resources.offheap.memory.mb": 0,
+        "storm.messaging.netty.max_wait_ms": 1000,
+        "pacemaker.port": 6699,
+        "task.heartbeat.frequency.secs": 3,
+        "storm.exhibitor.port": 8080,
+        "topology.metrics.consumer.register": [
+            {
+                "parallelism.hint": 2,
+                "class": "com.accelerate_experience.storm.metrics.statsd.StatsdMetricConsumer",
+                "argument": {
+                    "metrics.statsd.prefix": "storm.metrics.",
+                    "metrics.statsd.port": 8125,
+                    "metrics.statsd.host": "127.0.0.1"
+                }
+            }
+        ],
+        "task.refresh.poll.secs": 10,
+        "supervisor.blobstore.download.max_retries": 3,
+        "storm.workers.artifacts.dir": "workers-artifacts",
+        "drpc.https.port": -1,
+        "topology.tick.tuple.freq.secs": 4,
+        "topology.submitter.user": "storm",
+        "storm.zookeeper.root": "/storm",
+        "ui.http.creds.plugin": "org.apache.storm.security.auth.DefaultHttpCredentialsPlugin",
+        "storm.log4j2.conf.dir": "log4j2",
+        "worker.heartbeat.frequency.secs": 1,
+        "storm.cluster.state.store": "org.apache.storm.cluster_state.zookeeper_state_factory",
+        "storm.messaging.netty.buffer_size": 5242880,
+        "storm.local.mode.zmq": False,
+        "nimbus.task.launch.secs": 120,
+        "topology.users": [],
+        "backpressure.disruptor.low.watermark": 0.4,
+        "topology.executor.receive.buffer.size": 1024,
+        "worker.profiler.childopts": "-XX:+UnlockCommercialFeatures -XX:+FlightRecorder",
+        "nimbus.file.copy.expiration.secs": 600,
+        "drpc.authorizer.acl.strict": False,
+        "topology.worker.shared.thread.pool.size": 4,
+        "storm.health.check.dir": "/var/lib/storm/healthchecks",
+        "topology.transfer.buffer.size": 1024,
+        "supervisor.slots.ports": [
+            6700
+        ],
+        "topology.state.checkpoint.interval.ms": 1000,
+        "topology.worker.receiver.thread.count": 1,
+        "drpc.https.keystore.type": "JKS",
+        "task.credentials.poll.secs": 30,
+        "pacemaker.thread.timeout": 10,
+        "drpc.max_buffer_size": 1048576,
+        "transactional.zookeeper.port": None,
+        "dev.zookeeper.path": "/tmp/dev-storm-zookeeper",
+        "nimbus.inbox.jar.expiration.secs": 3600,
+        "storm.nimbus.retry.interval.millis": 2000,
+        "topology.submitter.principal": "",
+        "ui.host": "0.0.0.0",
+        "topology.spout.wait.strategy": "org.apache.storm.spout.SleepSpoutWaitStrategy",
+        "topology.worker.logwriter.childopts": "-Xmx64m",
+        "storm.daemon.metrics.reporter.plugins": [
+            "org.apache.storm.daemon.metrics.reporters.JmxPreparableReporter"
+        ],
+        "pacemaker.auth.method": "NONE",
+        "resource.aware.scheduler.priority.strategy": "org.apache.storm.scheduler.resource.strategies.priority.DefaultSchedulingPriorityStrategy",
+        "topology.executor.send.buffer.size": 1024,
+        "topology.scheduler.strategy": "org.apache.storm.scheduler.resource.strategies.scheduling.DefaultResourceAwareStrategy",
+        "logviewer.port": 9006,
+        "nimbus.code.sync.freq.secs": 120,
+        "drpc.https.keystore.password": "",
+        "topology.shellbolt.max.pending": 100,
+        "storm.blobstore.inputstream.buffer.size.bytes": 65536,
+        "supervisor.blobstore.class": "org.apache.storm.blobstore.NimbusBlobStore",
+        "topology.backpressure.enable": False,
+        "drpc.queue.size": 128,
+        "task.backpressure.poll.secs": 30,
+        "supervisor.blobstore.download.thread.count": 5,
+        "drpc.worker.threads": 64,
+        "supervisor.cpu.capacity": 300,
+        "topology.enable.message.timeouts": True,
+        "supervisor.heartbeat.frequency.secs": 5,
+        "storm.zookeeper.port": 2181,
+        "worker.log.level.reset.poll.secs": 30,
+        "storm.messaging.netty.min_wait_ms": 100,
+        "topology.stats.sample.rate": 1,
+        "supervisor.enable": True,
+        "zmq.linger.millis": 5000,
+        "topology.max.replication.wait.time.sec": 60,
+        "scheduler.display.resource": False,
+        "topology.sleep.spout.wait.strategy.time.ms": 1,
+        "logviewer.cleanup.interval.secsInterval": 86400,
+        "transactional.zookeeper.root": "/transactional",
+        "storm.group.mapping.service": "org.apache.storm.security.auth.ShellBasedGroupsMapping",
+        "zmq.threads": 1,
+        "topology.priority": 29,
+        "topology.builtin.metrics.bucket.size.secs": 60,
+        "nimbus.childopts": "-Xmx2048m",
+        "ui.filter.params": None,
+        "storm.cluster.mode": "distributed",
+        "storm.messaging.netty.client_worker_threads": 1,
+        "logviewer.max.per.worker.logs.size.mb": 2048,
+        "supervisor.run.worker.as.user": False,
+        "topology.max.task.parallelism": None,
+        "drpc.invocations.port": 3773,
+        "supervisor.localizer.cache.target.size.mb": 10240,
+        "topology.multilang.serializer": "org.apache.storm.multilang.JsonSerializer",
+        "storm.messaging.netty.server_worker_threads": 1,
+        "nimbus.blobstore.class": "org.apache.storm.blobstore.LocalFsBlobStore",
+        "resource.aware.scheduler.eviction.strategy": "org.apache.storm.scheduler.resource.strategies.eviction.DefaultEvictionStrategy",
+        "topology.max.error.report.per.interval": 5,
+        "storm.thrift.transport": "org.apache.storm.security.auth.SimpleTransportPlugin",
+        "zmq.hwm": 0,
+        "storm.group.mapping.service.params": None,
+        "worker.profiler.enabled": False,
+        "storm.principal.tolocal": "org.apache.storm.security.auth.DefaultPrincipalToLocal",
+        "supervisor.worker.shutdown.sleep.secs": 1,
+        "pacemaker.host": "localhost",
+        "storm.zookeeper.retry.times": 5,
+        "ui.actions.enabled": True,
+        "topology.acker.executors": None,
+        "topology.fall.back.on.java.serialization": True,
+        "topology.eventlogger.executors": 0,
+        "supervisor.localizer.cleanup.interval.ms": 600000,
+        "storm.zookeeper.servers": [
+            "zookeeper01.example.com"
+        ],
+        "nimbus.thrift.threads": 64,
+        "logviewer.cleanup.age.mins": 10080,
+        "topology.worker.childopts": None,
+        "topology.classpath": None,
+        "supervisor.monitor.frequency.secs": 3,
+        "nimbus.credential.renewers.freq.secs": 600,
+        "topology.skip.missing.kryo.registrations": False,
+        "drpc.authorizer.acl.filename": "drpc-auth-acl.yaml",
+        "pacemaker.kerberos.users": [],
+        "storm.group.mapping.service.cache.duration.secs": 120,
+        "topology.testing.always.try.serialize": False,
+        "nimbus.monitor.freq.secs": 10,
+        "worker.childops": "-Xmx2048m -XX:+PrintGCDetails -Xloggc:artifacts/gc.log -XX:+PrintGCDateStamps -XX:+PrintGCTimeStamps -XX:+UseGCLogFileRotation -XX:NumberOfGCLogFiles=10 -XX:GCLogFileSize=1M -XX:+HeapDumpOnOutOfMemoryError -XX:HeapDumpPath=artifacts/heapdump",
+        "storm.health.check.timeout.ms": 10000,
+        "supervisor.supervisors": [],
+        "topology.tasks": None,
+        "topology.bolts.outgoing.overflow.buffer.enable": False,
+        "storm.messaging.netty.socket.backlog": 500,
+        "topology.workers": 6,
+        "pacemaker.base.threads": 10,
+        "storm.local.dir": "/var/lib/storm/data",
+        "worker.childopts": "-Xmx%HEAP-MEM%m -XX:+PrintGCDetails -Xloggc:artifacts/gc.log -XX:+PrintGCDateStamps -XX:+PrintGCTimeStamps -XX:+UseGCLogFileRotation -XX:NumberOfGCLogFiles=10 -XX:GCLogFileSize=1M -XX:+HeapDumpOnOutOfMemoryError -XX:HeapDumpPath=artifacts/heapdump",
+        "storm.auth.simple-white-list.users": [],
+        "topology.disruptor.batch.timeout.millis": 1,
+        "topology.message.timeout.secs": 300,
+        "topology.state.synchronization.timeout.secs": 60,
+        "topology.tuple.serializer": "org.apache.storm.serialization.types.ListDelegateSerializer",
+        "supervisor.supervisors.commands": [],
+        "nimbus.blobstore.expiration.secs": 600,
+        "logviewer.childopts": "-Xmx128m",
+        "topology.environment": {
+            "mytopology.foo": "bar"
+        },
+        "topology.debug": False,
+        "topology.disruptor.batch.size": 100,
+        "storm.messaging.netty.max_retries": 300,
+        "ui.childopts": "-Xmx768m",
+        "storm.network.topography.plugin": "org.apache.storm.networktopography.DefaultRackDNSToSwitchMapping",
+        "storm.zookeeper.session.timeout": 20000,
+        "drpc.childopts": "-Xmx768m",
+        "drpc.http.creds.plugin": "org.apache.storm.security.auth.DefaultHttpCredentialsPlugin",
+        "storm.zookeeper.connection.timeout": 15000,
+        "storm.zookeeper.auth.user": None,
+        "storm.meta.serialization.delegate": "org.apache.storm.serialization.GzipThriftSerializationDelegate",
+        "topology.max.spout.pending": 500,
+        "storm.codedistributor.class": "org.apache.storm.codedistributor.LocalFileSystemCodeDistributor",
+        "nimbus.supervisor.timeout.secs": 60,
+        "nimbus.task.timeout.secs": 30,
+        "storm.zookeeper.superACL": None,
+        "drpc.port": 3772,
+        "pacemaker.max.threads": 50,
+        "storm.zookeeper.retry.intervalceiling.millis": 30000,
+        "nimbus.thrift.port": 6627,
+        "storm.auth.simple-acl.admins": [],
+        "topology.component.cpu.pcore.percent": 10,
+        "supervisor.memory.capacity.mb": 3072,
+        "storm.nimbus.retry.times": 5,
+        "supervisor.worker.start.timeout.secs": 120,
+        "storm.zookeeper.retry.interval": 1000,
+        "logs.users": None,
+        "worker.profiler.command": "flight.bash"
+    },
+    "uptime": "17d 15h 49m 48s",
+    "schedulerInfo": None,
+    "name": "my_topology",
+    "workersTotal": 6
+}
+
+TEST_STORM_TOPOLOGY_METRICS_RESP = {
+    "window": ":all-time",
+    "window-hint": "All time",
+    "spouts": [
+        {
+            "id": "spout",
+            "emitted": [
+                {"stream_id": "__metrics", "value": 20},
+                {"stream_id": "default", "value": 17350280},
+                {"stream_id": "__ack_init", "value": 17328160},
+                {"stream_id": "__system", "value": 20}
+            ],
+            "transferred": [
+                {"stream_id": "__metrics", "value": 20},
+                {"stream_id": "default", "value": 17350280},
+                {"stream_id": "__ack_init", "value": 17328160},
+                {"stream_id": "__system", "value": 0}
+            ],
+            "acked": [
+                {"stream_id": "default", "value": 17339180}
+            ],
+            "failed": [],
+            "complete_ms_avg": [
+                {"stream_id": "default", "value": "920.497"}
+            ]
+        }
+    ],
+    "bolts": [
+        {
+            "id": "count",
+            "emitted": [
+                {"stream_id": "__metrics", "value": 120},
+                {"stream_id": "default", "value": 190748180},
+                {"stream_id": "__ack_ack", "value": 190718100},
+                {"stream_id": "__system", "value": 20}
+            ],
+            "transferred": [
+                {"stream_id": "__metrics", "value": 120},
+                {"stream_id": "default", "value": 0},
+                {"stream_id": "__ack_ack", "value": 190718100},
+                {"stream_id": "__system", "value": 0}
+            ],
+            "acked": [
+                {"component_id": "split", "stream_id": "default", "value": 190733160}
+            ],
+            "failed": [],
+            "process_ms_avg": [
+                {"component_id": "split", "stream_id": "default", "value": "0.004"}
+            ],
+            "executed": [
+                {"component_id": "split", "stream_id": "default", "value": 190733140}
+            ],
+            "executed_ms_avg": [
+                {"component_id": "split", "stream_id": "default", "value": "0.005"}
+            ]
+        },
+        {
+            "id": "split",
+            "emitted": [
+                {"stream_id": "__metrics", "value": 60},
+                {"stream_id": "default", "value": 190754740},
+                {"stream_id": "__ack_ack", "value": 17317580},
+                {"stream_id": "__system", "value": 20}
+            ],
+            "transferred": [
+                {"stream_id": "__metrics", "value": 60},
+                {"stream_id": "default", "value": 190754740},
+                {"stream_id": "__ack_ack", "value": 17317580},
+                {"stream_id": "__system", "value": 0}
+            ],
+            "acked": [
+                {"component_id": "spout", "stream_id": "default", "value": 17339180}
+            ],
+            "failed": [],
+            "process_ms_avg": [
+                {"component_id": "spout", "stream_id": "default", "value": "0.051"}
+            ],
+            "executed": [
+                {"component_id": "spout", "stream_id": "default", "value": 17339240}
+            ],
+            "executed_ms_avg": [
+                {"component_id": "spout", "stream_id": "default", "value": "0.052"}
+            ]
+        }
+    ]
+}
