+++ storm/tests/compose/Dockerfile
