+++ storm/tox.ini
@@ -1,0 +1,29 @@
+[tox]
+minversion = 2.0
+skip_missing_interpreters = true
+basepython = py37
+envlist =
+    py{27,37}-storm
+    flake8
+
+[testenv]
+usedevelop = true
+platform = linux|darwin|win32
+deps =
+    datadog-checks-base[deps]
+    -rrequirements-dev.txt
+passenv =
+    DOCKER*
+    COMPOSE*
+commands =
+    pip install -r requirements.in
+    pytest -v
+
+[testenv:flake8]
+skip_install = true
+deps = flake8
+commands = flake8 .
+
+[flake8]
+exclude = .eggs,.tox,build
+max-line-length = 120
